pyproject.toml
[tool.poetry]
name = "llm-test-assignment"
version = "0.1.0"
description = ""
authors = ["Igor Emtsev"]
license = "GPL"
readme = "README.md"
packages = [{include = "llm_test_assignment"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
langchain = "^0.3.11"
openai = "^1.57.2"
faiss-cpu = "^1.9.0.post1"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.11"
tiktoken = "^0.8.0"
response = "^0.5.0"
pytest = "^8.3.4"
sympy = "^1.13.3"
networkx = "^3.4.2"
sentence-transformers = "^3.3.1"
bert-score = "^0.3.13"
allure-pytest = "^2.13.5"
pytest-html = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
